+++ test/integration_test.jl
@@ -14,6 +14,8 @@ import ClimaComms
     ClimaComms.@import_required_backends
 end
 
+import LazyBroadcast: @lazy
+
 const context = ClimaComms.context()
 ClimaComms.init(context)
 
@@ -52,10 +54,30 @@ function setup_integrator(output_dir; context, more_compute_diagnostics = 0)
         end
     end
 
+    function compute_my_var_lazy(u, p, t)
+        return @lazy @. u.my_var + 10.0
+    end
+
+    function compute_my_var_field(u, p, t)
+        return u.my_var
+    end
+
     simple_var = ClimaDiagnostics.DiagnosticVariable(;
         compute! = compute_my_var!,
         short_name = "YO",
         long_name = "YO YO",
+    )
+
+    simple_var_lazy = ClimaDiagnostics.DiagnosticVariable(;
+        compute = compute_my_var_lazy,
+        short_name = "YO LAZY",
+        long_name = "YO YO LAZY",
+    )
+
+    simple_var_field = ClimaDiagnostics.DiagnosticVariable(;
+        compute = compute_my_var_field,
+        short_name = "YO LAZY FIELD",
+        long_name = "YO YO LAZY FIELD",
     )
 
     average_diagnostic = ClimaDiagnostics.ScheduledDiagnostic(
@@ -67,6 +89,14 @@ function setup_integrator(output_dir; context, more_compute_diagnostics = 0)
     )
     inst_diagnostic = ClimaDiagnostics.ScheduledDiagnostic(
         variable = simple_var,
+        output_writer = nc_writer,
+    )
+    inst_diagnostic_lazy = ClimaDiagnostics.ScheduledDiagnostic(
+        variable = simple_var_lazy,
+        output_writer = nc_writer,
+    )
+    inst_diagnostic_field = ClimaDiagnostics.ScheduledDiagnostic(
+        variable = simple_var_field,
         output_writer = nc_writer,
     )
     inst_every3s_diagnostic = ClimaDiagnostics.ScheduledDiagnostic(
@@ -92,6 +122,8 @@ function setup_integrator(output_dir; context, more_compute_diagnostics = 0)
     scheduled_diagnostics = [
         average_diagnostic,
         inst_diagnostic,
+        inst_diagnostic_lazy,
+        inst_diagnostic_field,
         inst_diagnostic_h5,
         inst_every3s_diagnostic,
         inst_every3s_diagnostic_another,